Finished workout CRUD with JWT authentication

+++ src/tasks/controller.py
@@ -1,39 +1,53 @@
 from src.tasks.dtos import Workout_Schema
 from sqlalchemy.orm import Session
 from src.tasks.models import WorkoutModel
-from fastapi import HTTPException
+from fastapi import HTTPException, Response, status
+from src.user.models import UserModel
 
-def add_workout(body: Workout_Schema, db: Session):
+
+def add_workout(body: Workout_Schema, db: Session, user: UserModel):
   data = body.model_dump()
   #new_workout = WorkoutModel(title = data["title"], description = data["description"], is_finished = data["is_finished"])
   # This does the EXACT SAME THING as your 3 lines of code:
   new_workout = WorkoutModel(**data)
+  new_workout.user_id = user.id
   db.add(new_workout)
   db.commit()
   db.refresh(new_workout)
   
-  return {
-    "status": "workout created successfully", "data": new_workout
-  }
+  return new_workout
 
-def get_all_workouts(db: Session):
-  all_workouts = db.query(WorkoutModel).all()
-  return {"status": "workouts retrieved successfully", "data": all_workouts}
+def get_all_workouts(db: Session, user: UserModel):
+  all_workouts = db.query(WorkoutModel).filter(WorkoutModel.user_id == user.id).all()
+  return all_workouts
 
-def get_workout(workout_id: int, db: Session):
+def get_workout(workout_id: int, db: Session, user: UserModel):
   one_workout =  db.query(WorkoutModel).filter(WorkoutModel.id == workout_id).first()
   if not one_workout:
     raise HTTPException(404, detail="Workout ID is incorrect")
   
-  return {"status": "workout retrieved successfully", "data": one_workout}
+  if one_workout.user_id != user.id:
+    raise HTTPException(403, detail="You are not authorized to view this workout")
+  
+  return one_workout
 
-def get_workout_type(workout_type: str, db: Session):
-  workout_type = db.query(WorkoutModel).filter(WorkoutModel.workout_type == workout_type).first() 
+def get_workout_type(workout_type: str, db: Session, user: UserModel):
+  workout = db.query(WorkoutModel).filter(WorkoutModel.workout_type == workout_type, WorkoutModel.user_id == user.id).first() 
+  if not workout:
+    raise HTTPException(404, detail="Workout type not found")
+  
+  if workout.user_id != user.id:
+    raise HTTPException(403, detail="You are not authorized to view this workout type")
+  
+  return workout
 
-def update_workout(body: Workout_Schema, workout_id: int, db: Session):
+def update_workout(body: Workout_Schema, workout_id: int, db: Session, user: UserModel):
   workout = db.query(WorkoutModel).filter(WorkoutModel.id == workout_id).first()
   if not workout:
     raise HTTPException(404, "Workout ID is incorrect")
+  
+  if workout.user_id != user.id:
+    raise HTTPException(403, "You are not authorized to update this workout")
   
   body = body.model_dump()
   for key, value in body.items():
@@ -43,17 +57,20 @@ def update_workout(body: Workout_Schema, workout_id: int, db: Session):
   db.commit()
   db.refresh(workout)
 
-  return {"status": "Workout updated successfully", "data": workout}
+  return workout
 
-def delete_workout(workout_id: int, db: Session):
+def delete_workout(workout_id: int, db: Session, user: UserModel):
   workout = db.query(WorkoutModel).get(workout_id)
   if not workout:
     raise HTTPException(404, "Workout ID not found")
   
+  if workout.user_id != user.id:
+    raise HTTPException(403, "You are not authorized to delete this workout")
+  
   db.delete(workout)
   db.commit()
 
-  return None
+  return Response(status_code=status.HTTP_204_NO_CONTENT)
